rename topology modules
